move ModeView from components
